[Pipelining] Relax scale_grads assert
The assert is morally valid- if no gradients are scaled, then something
is definitely wrong with the setup.  In one instance, PP +
optimizer-in-backward (in torchtitan) resulted in grad=None after
running .backward() and before scaling grads.  This is obviously not a
correct scenario.

On the other hand, the existing assert is too restrictive.  It's
possible that a model used with pipelining would have some parameters
that do not receieve gradients, and we shouldn't hard-error in these
cases.  (E.g. if the parameter is literally not used, or is frozen).

ghstack-source-id: f2e7e51d192768296c6acfba3362123f5a9e3c25
Pull Request resolved: https://github.com/pytorch/pytorch/pull/145010